# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API DANAZ Pay",
        "description": "API pour la gestion des paiements CNPS, Banques et Entreprises.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Principal local"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connecter un utilisateur (Banque, CNPS ou Superviseur)",
                "description": "Permet à un agent CNPS, une banque ou un superviseur de se connecter et de récupérer un token Sanctum.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "guichet@banque.cm"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "authentification reussie"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abcdef123456..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/login-company": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion d'une entreprise via Numéro Employeur",
                "description": "Authentifie ou crée une entreprise avec son Numéro Employeur après vérification sur la plateforme CNPS, et génère un token Danazpay.",
                "operationId": "loginCompany",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_employeur"
                                ],
                                "properties": {
                                    "numero_employeur": {
                                        "description": "Numéro Employeur CNPS",
                                        "type": "string",
                                        "example": "M0123456789"
                                    },
                                    "name": {
                                        "description": "Raison sociale (optionnelle)",
                                        "type": "string",
                                        "example": "Ikarootech"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Authentification réussie"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abcdef..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Numéro employeur non reconnu par la CNPS"
                    },
                    "503": {
                        "description": "Service CNPS indisponible"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Récupérer le profil de l'utilisateur connecté",
                "description": "Retourne les informations de l'utilisateur actuel. Nécessite le token Bearer.",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Profil récupéré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié (Token manquant ou invalide)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "description": "Détruit le token actuel de l'utilisateur.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}